backend/requirements.txt
africastalking==2.0.1
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.0.14
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-decouple==3.8
PyYAML==6.0.3
requests==2.32.5
responses==0.25.8
schema==0.7.8
sqlparse==0.5.5
urllib3==2.6.2
